Psalm fix

+++ src/Blog/Application/Service/AppService/CommandService/AuthorPostService.php
@@ -8,11 +8,13 @@ use App\Blog\Application\Service\CommandService\AuthorPostServiceInterface;
 use App\Blog\Application\Service\CommandService\PostChangeDTO;
 use App\Blog\Application\Service\CommandService\PostCreateDTO;
 use App\Blog\Application\Service\QueryService\AuthorPostQueryServiceInterface;
+use App\Blog\Domain\Exception\BlogAccessDeniedException;
 use App\Blog\Domain\Exception\BlogNotFoundException;
 use App\Blog\Domain\Port\PostRepositoryInterface;
 use App\Blog\Domain\Port\TagRepositoryInterface;
 use App\Blog\Domain\Post;
 use App\Blog\Domain\User\Author;
+use Override;
 
 final readonly class AuthorPostService implements AuthorPostServiceInterface
 {
@@ -23,7 +25,7 @@ final readonly class AuthorPostService implements AuthorPostServiceInterface
     ) {
     }
 
-    #[\Override]
+    #[Override]
     public function create(PostCreateDTO $postCreateDTO, Author $author): void
     {
         $post = new Post(
@@ -41,13 +43,22 @@ final readonly class AuthorPostService implements AuthorPostServiceInterface
 
     /**
      * @throws BlogNotFoundException
+     * @throws BlogAccessDeniedException
      */
-    #[\Override]
-    public function edit(string $postSlug, PostChangeDTO $postChangeDTO): void
-    {
+    #[Override]
+    public function edit(
+        string $postSlug,
+        PostChangeDTO $postChangeDTO,
+        Author $author,
+    ): void {
         if (($post = $this->postQueryService->getPostBySlug($postSlug)) === null) {
             throw new BlogNotFoundException('This post does not exist!');
         }
+
+        if ($post->getAuthor() !== $author) {
+            throw new BlogAccessDeniedException('You are not the author of this post!');
+        }
+
         $post->edit($postChangeDTO->getTitle(), $postChangeDTO->getContent());
         $post->resetTags();
 
@@ -61,13 +72,19 @@ final readonly class AuthorPostService implements AuthorPostServiceInterface
 
     /**
      * @throws BlogNotFoundException
+     * @throws BlogAccessDeniedException
      */
-    #[\Override]
-    public function delete(string $postSlug): void
+    #[Override]
+    public function delete(string $postSlug, Author $author): void
     {
         if (($post = $this->postQueryService->getPostBySlug($postSlug)) === null) {
             throw new BlogNotFoundException('This post does not exist!');
         }
+
+        if (!$post->isAuthor($author)) {
+            throw new BlogAccessDeniedException('You are not the author of this post!');
+        }
+
         $this->repository->delete([$post]);
     }
